+++ tests/Rector/Misc/PropertyToConfigRector/config/configured_rule.php
@@ -7,7 +7,7 @@ use Rector\Config\RectorConfig;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->ruleWithConfiguration(PropertyToConfigRector::class, [
-        'SilverStripe\Dev\BuildTask' => [
+        \SilverStripe\Dev\BuildTask::class => [
             'enabled' => [
                 'name' => 'is_enabled',
                 'visibility' => \PhpParser\Modifiers::PRIVATE | \PhpParser\Modifiers::STATIC,