+++ tests/Rector/Misc/StaticCallToConfigRector/config/configured_rule.php
@@ -7,19 +7,19 @@ use Rector\Config\RectorConfig;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->ruleWithConfiguration(StaticCallToConfigRector::class, [
-        'SilverStripe\ORM\DataObject' => [
+        \SilverStripe\ORM\DataObject::class => [
             'disable_subclass_access' => [
                 'property' => 'subclass_access',
                 'value' => false,
                 'visibility' => \PhpParser\Modifiers::PRIVATE | \PhpParser\Modifiers::STATIC,
-                'add_config' => true,
+                'add_config' => true
             ],
             'enable_subclass_access' => [
                 'property' => 'subclass_access',
                 'value' => true,
                 'visibility' => \PhpParser\Modifiers::PRIVATE | \PhpParser\Modifiers::STATIC,
-                'add_config' => true,
-            ],
-        ],
+                'add_config' => true
+            ]
+        ]
     ]);
 };